+++ CHANGELOG.md
@@ -11,11 +11,14 @@
 
 # Release Notes
 
-## [Vulkan Profiles Toolset 1.3.XXX](https://github.com/KhronosGroup/Vulkan-Profiles/tree/main) - XXX 2023
+## [Vulkan Profiles Toolset 1.3.250](https://github.com/KhronosGroup/Vulkan-Profiles/tree/main) - XXX 2023
 
 ### Features:
 - Add a setting to control how layer features are initialized when a feature is not specified in a selected profile.
 - Add Profiles layer loading of profiles with multiple variants. All variants are emulated.
+
+### Improvements:
+- Indicate the corresponding device name to log message when a profile capability is not supported by the given device.
 
 ### Bugfixes:
 - Fix invalid message when trying to override a supported capability to unsupported #393